API Reference

GeNIOS.Optimizer

GeNIOS.generate_random_sdp

GeNIOS.unvec_symm

GeNIOS.vec_symm

GeNIOS.Optimizer — Type

Optimizer()

Create a new GeNIOS optimizer.

GeNIOS.generate_random_sdp — Method

generate_random_sdp(n; rand_seed=0)

Generates a random dual form SDP with side dimension n: min c'*x s.t. sum(F[i]*x[i]) + G ⪰ 0

Returns c, F, G, xstar, D, where xstar and D are optimal primal and dual variables respectively

GeNIOS.unvec_symm — Method

unvec_symm(x)

Returns a dim-by-dim symmetric matrix corresponding to x. x is a vector of length dim*(dim + 1)/2, corresponding to a symmetric matrix

X = [ X11     X12/√2 ... X1k/√2
      X21/√2  X22    ... X2k/√2
      ...
      Xk1/√2  Xk2/√2 ... Xkk ],

where vec(X = (X11, X12, X22, X13, X23, ..., Xkk)

Note that the factor √2 preserves inner products: x'*c = Tr(unvec_symm(c, dim) * unvec_symm(x, dim))

GeNIOS.vec_symm — Method

vec_symm(X)

Returns a vectorized representation of a symmetric matrix X. vec(X) = (X11, √2*X12, X22, √2*X13, X23, ..., Xkk)

Note that the factor √2 preserves inner products: x'*c = Tr(unvec_symm(c, dim) * unvec_symm(x, dim))